+++ src/game.c
@@ -24,7 +24,7 @@
 #include "audio.h"
 #include "unknown_00BC20.h"
 #include "camera.h"
-#include "controller_pak.h"
+#include "save_data.h"
 #include "unknown_078050.h"
 #include "unknown_003260.h"
 #include "unknown_032760.h"
