+++ src/menu.c
@@ -23,7 +23,7 @@
 #include "unknown_00BC20.h"
 #include "game.h"
 #include "unknown_078050.h"
-#include "controller_pak.h"
+#include "save_data.h"
 #include "object_functions.h"
 #include "unknown_003260.h"
