+++ src/racer.c
@@ -10,7 +10,7 @@
 #include "macros.h"
 #include "structs.h"
 #include "asset_sections.h"
-#include "controller_pak.h"
+#include "save_data.h"
 #include "unknown_008C40.h"
 #include "unknown_00BC20.h"
 #include "audio.h"
